backend/requirements.txt
fastapi==0.124.4
sqlmodel==0.0.31
pg8000==1.30.3
pyjwt==2.10.1
python-multipart==0.0.21
python-dotenv==1.0.0
uvicorn==0.38.0
alembic==1.17.2
pydantic-settings==2.1.0